Initial window creation.
This is the (most) naive approach, providing a single endpoint that is
intended to be used to create any/all window types. The problem with
this approach is that the way FastAPI uses the annotations in the
signature of a method, just using the basic WindowParams and Window
types means that the request data and response data is "trimmed" to
include only the fields in those types, and not make use of the fields
in other types.